+++ src/modules/posts/posts.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from "@nestjs/common";
 import { InjectModel } from "@nestjs/sequelize";
 
 import { Post } from "./models/post.model";
-import { Tag } from "./models/tags.model";
+import { Tag } from "./models/tag.model";
 import { Users } from "db/models/users.model";
 
 @Injectable()
